fix exceptions
