+++ akka-http-jwt/src/main/resources/reference.conf
