+++ akka-http-jwt/src/main/scala/akka/http/jwt/impl/settings/JwtKeyLoader.scala
@@ -1,10 +1,7 @@
-/*
- * Copyright 2022 Lightbend Inc.
- */
+package akka.http.jwt.impl.settings
 
-package akka.http.jwt.util
-
-import akka.http.jwt.util.JwtSupport.{ JwtAsymmetricAlgorithmSecret, JwtNoneAlgorithmSecret }
+import akka.http.jwt.impl.settings.JwtSupport.JwtAsymmetricAlgorithmSecret
+import akka.pki.pem.{ DERPrivateKeyLoader, PEMDecoder, PEMLoadingException }
 import pdi.jwt.{ JwtAlgorithm, algorithms }
 
 import java.io.File
@@ -12,7 +9,6 @@ import java.nio.file.Files
 import java.security.spec.{ PKCS8EncodedKeySpec, X509EncodedKeySpec }
 import java.security.{ KeyFactory, KeyPair }
 import javax.crypto.spec.SecretKeySpec
-import akka.pki.pem.{ DERPrivateKeyLoader, PEMDecoder, PEMLoadingException }
 
 object JwtKeyLoader {
 
@@ -25,7 +21,7 @@ object JwtKeyLoader {
             symmetric,
             new SecretKeySpec(Files.readAllBytes(secretKeyFile.toPath), algorithm.fullName))
         } else {
-          throw new RuntimeException("Jwt.expectedHmacSecret(keyId)")
+          throw new IllegalArgumentException(s"Expected a symmetric secret configured for JWT key with id [$keyId]")
         }
       case asymmetric: algorithms.JwtAsymmetricAlgorithm =>
         val keyAlgo = asymmetric match {
@@ -43,9 +39,9 @@ object JwtKeyLoader {
               try {
                 Some(KeyFactory.getInstance(keyAlgo).generatePublic(new X509EncodedKeySpec(pem.bytes)))
               } catch {
-                case e: Exception => throw new RuntimeException("Jwt.keyDecodingError(keyId, public, e.getMessage)")
+                case e: Exception => throw new IllegalArgumentException(s"Error decoding JWT public key from key id [$keyId]: ${e.getMessage}", e)
               }
-            case _ => throw new RuntimeException("Jwt.unknownPublicKeyType(keyId, pem.label)")
+            case _ => throw new IllegalArgumentException(s"Unsupported JWT public key format for key id [$keyId]: ${pem.label}")
           }
         } else None
 
@@ -59,32 +55,34 @@ object JwtKeyLoader {
               try {
                 Some(KeyFactory.getInstance(keyAlgo).generatePrivate(new PKCS8EncodedKeySpec(pem.bytes)))
               } catch {
-                case e: Exception => throw new RuntimeException("Jwt.keyDecodingError(keyId, private, e.getMessage)")
+                case e: Exception => throw new IllegalArgumentException(
+                  s"Error decoding JWT private key from key id [$keyId]: ${e.getMessage}", e)
               }
 
             case "RSA PRIVATE KEY" =>
               try {
                 Some(DERPrivateKeyLoader.load(pem))
               } catch {
-                case e: Exception => throw new RuntimeException("Jwt.keyDecodingError(keyId, private, e.getMessage)")
+                case e: Exception => throw new IllegalArgumentException(
+                  s"Error decoding JWT private key from key id [$keyId]: ${e.getMessage}", e)
               }
 
             case "EC PRIVATE KEY" =>
-              throw new RuntimeException("Jwt.ecPrivateKeyNotSupported(keyId)")
+              throw new IllegalArgumentException(s"Raw ECDSA JWT private key in key id [$keyId] is not supported.")
 
             case _ =>
-              throw new RuntimeException("Jwt.unknownPrivateKeyType(keyId, pem.label)")
+              throw new IllegalArgumentException(s"Unsupported JWT private key format for key id [$keyId]: ${pem.label}")
           }
         } else None
 
         (privateKey, publicKey) match {
           case (None, None) =>
-            throw new RuntimeException("Jwt.expectedAsymmetricSecret(keyId)")
+            throw new IllegalArgumentException(s"Expected an asymmetric secret configured for JWT key with id [$keyId]")
           case (maybePriv, maybePub) =>
             JwtAsymmetricAlgorithmSecret(asymmetric, new KeyPair(maybePub.orNull, maybePriv.orNull))
         }
       case other =>
-        throw new RuntimeException("Jwt.unknownAlgorithm")
+        throw new IllegalArgumentException(s"Unknown JWT algorithm for key id [$keyId]: $other")
     }
 
   }
@@ -94,7 +92,7 @@ object JwtKeyLoader {
       PEMDecoder.decode(new String(Files.readAllBytes(file.toPath)))
     } catch {
       case e: PEMLoadingException =>
-        throw new RuntimeException("Jwt.pemDecodingError(keyId, name, e.getMessage)")
+        throw new IllegalArgumentException(s"Error PEM decoding JWT $name key from key id [$keyId]: ${e.getMessage}", e)
     }
   }
