+++ akka-http-jwt/src/main/scala/akka/http/jwt/impl/settings/JwtSprayJson.scala
@@ -1,14 +1,30 @@
 /*
- * Copyright 2022 Lightbend Inc.
+ * Copyright 2024 Lightbend Inc.
  */
 
-package akka.http.jwt.util
+/*
+ * Copied from github.com/jwt-scala/jwt-scala/ with this license:
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
+package akka.http.jwt.impl.settings
+
+import pdi.jwt.exceptions.JwtNonStringException
 import pdi.jwt.{ JwtAlgorithm, JwtClaim, JwtHeader, JwtJsonCommon }
+import spray.json._
 
 import java.time.Clock
-import pdi.jwt.exceptions.JwtNonStringException
-import spray.json._
 
 /**
  * Implementation of `JwtCore` using `JsObject` from spray-json.
